requirements.txt
# Web API
Flask==1.1.2
connexion[swagger-ui]==2.7.0

# Database
SQLAlchemy==1.3.20
alembic==1.4.3
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
psycopg2-binary==2.8.6

# Admin
Flask-Admin==1.5.7
Flask-BasicAuth==0.2.0

# AUTH
PyJWT==2.0.0
